Core cleanup (removed redundant core parts to compatibility)
