+++ core/compatibility/lib/themes.js
@@ -1,0 +1,26 @@
+import merge from 'lodash-es/merge'
+import * as corePlugins from '@vue-storefront/core/plugins'
+import * as coreMixins from '@vue-storefront/core/mixins'
+import * as coreFilters from '@vue-storefront/core/filters'
+import * as themeMixins from 'theme/mixins'
+import * as themeFilters from 'theme/filters'
+
+/** DEPRECIATED will be removed in 1.7 */
+export function plugins () {
+  return corePlugins
+}
+
+/** DEPRECIATED will be removed in 1.7 */
+export function mixins () {
+  return merge(coreMixins, themeMixins)
+}
+
+/** DEPRECIATED will be removed in 1.7 */
+export function filters () {
+  return merge(themeFilters, coreFilters)
+}
+
+/** DEPRECIATED will be removed in 1.7 */
+export function extendStore (coreStore, extendStore) {
+  return merge(coreStore, extendStore)
+}
